# api/docs/swagger.yaml
basePath: /v1
definitions:
  models.AdminLoginRes:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
  models.AdminReq:
    properties:
      password:
        type: string
      user_name:
        type: string
    type: object
  models.AdminRes:
    properties:
      access_token:
        type: string
      created_at:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      user_name:
        type: string
    type: object
  models.AllAdmins:
    properties:
      admins:
        items:
          $ref: '#/definitions/models.GetAdminProdile'
        type: array
    type: object
...
  models.BrandCategories:
    properties:
      category_name:
        type: string
      id:
        type: integer
    type: object
...
  models.BrandRequest:
    properties:
      brand_name:
        type: string
    type: object
  models.BrandResponse:
    properties:
      brand_name:
        type: string
      id:
        type: integer
      logo:
        type: string
    type: object
  models.BrandUpdateReq:
    properties:
      brand_name:
        type: string
      id:
        type: integer
    type: object
...
  models.FailureInfo:
    properties:
      code:
        type: integer
      error: {}
      message:
        type: string
    type: object
...
  models.GetAdminProdile:
    properties:
      created_at:
        type: string
      id:
        type: string
      password:
        type: string
      updated_at:
        type: string
      user_name:
        type: string
    type: object
  models.GetBrandInfo:
    properties:
      brand_name:
        type: string
      categories:
        items:
          $ref: '#/definitions/models.BrandCategories'
        type: array
      id:
        type: integer
    type: object
...
  models.SuccessInfo:
    properties:
      message:
        type: string
      status_code:
        type: integer
    type: object
...
host: 100.26.188.92:7070
info:
  contact:
    email: gofurovmurtazoxon@gmail.com
    name: Murtazoxon
    url: https://t.me/murtazokhon_gofurov
  description: This is climate_uz server api server
  termsOfService: 2 term climate_uz
  title: Climate_uz api
  version: "1.0"
paths:
  /admin:
    patch:
      consumes:
      - application/json
      description: Through this api admin can update info
      parameters:
      - description: Admin Update
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAdminProdile'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Update admin info
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Through this api new admin can register
      parameters:
      - description: AddAdmin
        in: body
        name: admin
        required: true
        schema:
          $ref: '#/definitions/models.AdminReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.AdminRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Add admin
      tags:
      - Admin
  /admin/{id}:
    delete:
      consumes:
      - application/json
      description: Through this api admin can delete info
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Delete admin info
      tags:
      - Admin
  /admin/{username}/{password}:
    get:
      consumes:
      - application/json
      description: Through this api admin can login
      parameters:
      - description: username
        in: path
        name: username
        required: true
        type: string
      - description: password
        in: path
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AdminLoginRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: Login admin
      tags:
      - Admin
  /admin/profile:
    get:
      consumes:
      - application/json
      description: Through this api admin can get profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetAdminProdile'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Get admin profile
      tags:
      - Admin
  /admins:
    get:
      consumes:
      - application/json
      description: Through this api admin can get all admins
      parameters:
      - description: search
        in: path
        name: keyword
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.AllAdmins'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Get all admin
      tags:
      - Admin
  /brand:
    patch:
      consumes:
      - application/json
      description: Through this api can update brand info
      parameters:
      - description: Brand Update
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/models.BrandUpdateReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.BrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Update brand info
      tags:
      - Brand
    post:
      consumes:
      - application/json
      description: This method add new brand
      parameters:
      - description: AddNewBrand
        in: body
        name: brand
        required: true
        schema:
          $ref: '#/definitions/models.BrandRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.BrandResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      security:
      - BearerAuth: []
      summary: Add new brand
      tags:
      - Brand
...
  /brand/{id}:
    get:
      consumes:
      - application/json
      description: Through this api can get brand info
      parameters:
      - description: id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.GetBrandInfo'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.FailureInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.FailureInfo'
      summary: Get brand info
      tags:
      - Brand
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"